src/test/java/com/rest/springbootemployee/EmployeeServiceTest.java
package com.rest.springbootemployee;


import com.rest.springbootemployee.pojo.Employee;
import com.rest.springbootemployee.repository.EmployeeRepository;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.springframework.test.context.junit.jupiter.SpringExtension;

import java.util.ArrayList;
import java.util.List;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.mockito.BDDMockito.given;

@ExtendWith(SpringExtension.class)
public class EmployeeServiceTest {

    @Mock
    EmployeeRepository employeeRepository;

    @InjectMocks
    EmployeeService employeeService;

    @Test
    void should_return_all_employees_when_find_all_given_employees() {
        //given
        ArrayList<Employee> preparedEmployees = new ArrayList<>();
        Employee firstEmployee = new Employee(1, "Susan", 23, "female", 10000);
        Employee secondEmployee = new Employee(1, "Mathew", 25, "male", 12000);
        preparedEmployees.add(firstEmployee);
        preparedEmployees.add(secondEmployee);
        given(employeeRepository.findAll()).willReturn(preparedEmployees);

        //when
        List<Employee> employees = employeeService.findAll();

        //then

        assertEquals(2, employees.size());
        assertEquals(firstEmployee, employees.get(0));
    }
}
